twitter-monitor-backend/requirements.txt
tweepy>=4.12.0
python-telegram-bot>=20.0
schedule>=1.1.0
requests>=2.28.0
flask>=2.0.1
fastapi==0.110.0
uvicorn==0.29.0
pydantic>=1.10.7
python-dotenv>=1.0.0